+++ crates/ruma-common/src/directory.rs
@@ -14,8 +14,8 @@ use crate::{
 /// A chunk of a room list response, describing one room.
 ///
 /// To create an instance of this type, first create a [`PublicRoomsChunkInit`] and convert it via
-/// `PublicRoomsChunk::from` / `.into()`. It is also possible to construct this type from a
-/// [`RoomSummary`].
+/// `PublicRoomsChunk::from` / `.into()`. It is also possible to construct this type from or convert
+/// it to a [`RoomSummary`].
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
 pub struct PublicRoomsChunk {
@@ -138,6 +138,38 @@ impl From<RoomSummary> for PublicRoomsChunk {
             avatar_url,
             join_rule: join_rule.as_str().into(),
             room_type,
+        }
+    }
+}
+
+impl From<PublicRoomsChunk> for RoomSummary {
+    fn from(value: PublicRoomsChunk) -> Self {
+        let PublicRoomsChunk {
+            room_id,
+            canonical_alias,
+            name,
+            topic,
+            avatar_url,
+            room_type,
+            num_joined_members,
+            join_rule,
+            world_readable,
+            guest_can_join,
+        } = value;
+
+        Self {
+            canonical_alias,
+            name,
+            num_joined_members,
+            room_id,
+            topic,
+            world_readable,
+            guest_can_join,
+            avatar_url,
+            join_rule: join_rule.into(),
+            room_type,
+            encryption: None,
+            room_version: None,
         }
     }
 }